+++ Source/Krypton Components/Krypton.Toolkit/Controls Visuals/VisualCustomFileDialogForm.Designer.cs
@@ -459,7 +459,7 @@
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(735, 552);
+            this.ClientSize = new System.Drawing.Size(946, 533);
             this.Controls.Add(this._rootPanel);
             this.Margin = new System.Windows.Forms.Padding(2);
             this.MinimumSize = new System.Drawing.Size(679, 495);